Dry refactor getHostObject and getServiceObject
